CRUD/clientes/requirements.txt
Flask==3.0.3
psycopg[binary]==3.2.9
python-dotenv==1.0.1
